[#140] [#143] Renamed a number of files to make their purpose clearer
